app/requirements-dev.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
docutils==0.20.1
coverage==7.3.2
django-debug-toolbar==4.2.0
flake8==6.1.0